Change ParseUevent to get serial number
