+++ solo/cli/_patches.py
@@ -101,23 +101,3 @@ if sys.platform.startswith("darwin"):
         return descriptors
 
     fido2._pyu2f.macos.MacOsHidDevice.Enumerate = newEnumerate
-
-
-## Linux
-if sys.platform.startswith("linux"):
-    import fido2._pyu2f.linux
-
-    oldnewParseUevent = fido2._pyu2f.linux.ParseUevent
-
-    def newParseUevent(uevent, desc):
-        oldnewParseUevent(uevent, desc)
-        lines = uevent.split(b"\n")
-        for line in lines:
-            line = line.strip()
-            if not line:
-                continue
-            k, v = line.split(b"=")
-            if k == b"HID_UNIQ":
-                desc.serial_number = v.decode("utf8")
-
-    fido2._pyu2f.linux.ParseUevent = newParseUevent
